requirements.txt
alembic==1.7.5
anyio==3.5.0

apispec==3.3.2
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
blinker==1.4
cattrs==1.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.15
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
dnspython==2.2.0
docutils==0.16
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.1.0

Flask-AppBuilder==3.4.3
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fuzzywuzzy==0.18.0
graphviz==0.19.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
inflection==0.5.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
iso8601==1.0.2
isodate==0.6.1
itsdangerous==1.1.0
jarowinkler==1.0.2
jedi==0.18.1
jellyfish==0.8.9

jinjasql
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.6
notebook==6.4.6
numpy==1.21.5
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.2
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
prison==0.2.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyaml==21.10.1
pycparser==2.21
Pygments==2.11.1
PyJWT==1.7.1
pyparsing==3.0.6
pyreadstat==1.1.3
pyrsistent==0.16.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
rapidfuzz==2.0.11
regex==2021.11.10
requests==2.27.1
rfc3986==1.5.0
rich==11.0.0
Send2Trash==1.8.0
setproctitle
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
unicodecsv==0.14.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.2
WTForms==2.3.3
zipp==3.7.0
functools==0.5.0
